==> scripts/03_check_if_on_cluster.Rmd <==
---
title: "check if some samples are already on the cluster"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(vdbR)
connect_database('~/dbConfig.txt')
get_table_from_database('asv_counts_ag', pre_filter = F)
get_table_from_database('asv_alpha_diversity_ag')
get_counts_subset( c('11.tp.02'))
get_table_from_database('asv_sequences_ag')
```

```{r}
cur <- read_delim('../data/ls_all_samps.txt', delim = ' ', skip = 1, col_names = F) %>% 
  filter(str_detect(X9, 'Sample'))

chi <- read_csv('../data/enterococcus_shotgun_lookup_list.csv')  %>% 
  mutate(sample_folder = str_extract(directory, 'Sample_.+$'))

length(intersect(cur$X9, chi$sample_folder))

setdiff(chi$sample_folder, cur$X9 )

chi %>% 
  select(fid) %>% 
  distinct(fid) %>% 
  write_csv('../data/chi_shotgun_list.csv', col_names = F)


task <- chi %>% 
  distinct(sampleid)

get_
```

```{r}
# the current shotgun lookup table 
source('~/db_connect_simple.R')
connect_database(config_file = '~/dbConfig.txt')
get_table_from_database('shotgun_lookup_ad')

# the ones in the doris project
doris <- shotgun_lookup_ad %>% 
  filter(str_detect(directory, 'ponced'))
```


# Marina's samples

```{r}
m <- read_csv('../data/Marina Extra shotgun samples for humann3- Doris.csv') %>% 
  distinct(`Samples that are in lookup_ag`)

# find out the full folder and the server of those samples 
marina <- shotgun_lookup_ad %>%
  distinct(sampleid, .keep_all = T) %>% 
  filter(sampleid %in% m$`Samples that are in lookup_ag`)  %>% 
  mutate(samplefolder = str_extract(directory, 'Sample_.+$'))

setdiff( m$`Samples that are in lookup_ag`, marina$sampleid)
  
# things that are currently on lilac
lilac <- read_delim('../data/samples.txt', delim = ' ', skip = 1, col_names = F)  %>% 
  mutate(fid = str_replace(X9, 'Sample_',''),
         fid = str_replace(fid, '_IGO_.+$',''))

length(intersect(lilac$fid, marina$fid))# already on lilac and also in marina list 

# the ones that need to be uploaded
new <- marina %>% 
  filter(!fid %in% lilac$fid)

new %>% 
  count(projectid)  
# Project_07051_C and Project_07976

# only marina
marina %>% 
  select(fid) %>% 
  write_csv('../data/only_marina.csv', col_names = F)

```
```{r}
# there are Oriana and Hana's samples need to be processed 2021-4-21
Oriana <- read_csv('~/Downloads/samples_humann3.csv')
Hana <- read_csv('~/Downloads/shotgun_old.csv')
two <- union(Oriana$sampleid, Hana$sampleid)
current <- read_delim('../data/current_samples.txt', delim = ' ', skip = 1, col_names = F)  %>% 
  mutate(fid = str_replace(X9, 'Sample_',''),
         fid = str_replace(fid, '_IGO_.+$',''))
two_df <- shotgun_lookup_ad %>% 
  filter(sampleid %in% two) %>% 
  distinct(sampleid, .keep_all = T)
length(intersect(two_df$fid, current$fid))

need <- two_df %>% 
  filter(!fid %in% current$fid) %>% 
  ungroup() %>% 
  count(projectid)

only <- shotgun_lookup_ad %>% 
  filter(sampleid %in% Oriana$sampleid) %>% 
  distinct(sampleid, .keep_all = T) %>% 
  select(fid)
only %>% 
  write_csv('../data/Oriana_old_only.csv', col_names = F)

```

```{r}
# gosh I need to figure out the samples that need to be run
hnew <- read_csv('~/Downloads/shotgun_old_all_samples.csv')
all4 <- union(two, hnew$sampleid)
all <- union(all4, marina$sampleid)
toberun <- shotgun_lookup_ad %>% 
  distinct(sampleid, .keep_all = T) %>% 
  filter(sampleid %in% all) 

toberun %>% 
  select(fid) %>% 
  write_csv('../data/new_huge.csv', col_names = F)
```

```{r}
#2021-08-31 for chi
# things that are currently on lilac
lilac <- read_delim('../../MSS_pipeline-/scripts/shotgun_pipeline/data/current.txt', delim = ' ', skip = 1, col_names = F)  %>% 
  mutate(fid = str_replace(X9, 'Sample_',''),
         fid = str_replace(fid, '_IGO_.+$',''))
chi <- read_csv('../../MSS_pipeline-/scripts/shotgun_pipeline/data/shotgun_fid_for_transition_state_project.csv') %>% 
  arrange(sampleid, projectid) %>% 
  distinct(sampleid, .keep_all = T)

length(intersect(lilac$fid, chi$fid))


miss <- setdiff(chi$fid,lilac$fid )

mis <- shotgun_lookup_ad %>% 
  filter(fid %in% miss)



all <- tibble(fid = intersect(lilac$fid, chi$fid))
split(all, (as.numeric(rownames(all))-1) %/% 300) %>% 
  imap(~ write_csv(x = .x, str_glue('../../MSS_pipeline-/scripts/shotgun_pipeline/data/chi{.y}.csv'), col_names = F))

```
